+++ src/logging/Log.ts
@@ -1,0 +1,11 @@
+import * as debug from 'debug';
+
+export function log(formatter: any, ...args: any[]) {
+    if (args && args.length) {
+        logger(formatter, args);
+    } else {
+        logger(formatter);
+    }
+}
+
+const logger = debug('axis-discovery-ssdp');