+++ src/root-descriptions/RootDescriptionRequest.ts
@@ -1,20 +1,17 @@
 import * as request from 'request';
 
+import { log } from './../logging/Log';
 import { RootDescription } from './RootDescription';
 
-/**
- * Class responsible for requesting a root description.
- */
 export class RootDescriptionRequest {
     constructor(
         private readonly remoteAddress: string,
         private readonly location: string) {
     }
 
-    /**
-     * Sends the request for a root description asynchronously.
-     */
     public async send(): Promise<RootDescription> {
+        log('RootDescriptionRequest#send - %s', this.remoteAddress);
+
         return new Promise<RootDescription>((resolve, reject) => {
             request.get(this.location, undefined, (error: any, _, body: string) => {
                 if (error) {