+++ test/network-interfaces/NetworkInterface.spec.ts
@@ -2,16 +2,16 @@ import * as chai from 'chai';
 import * as os from 'os';
 import * as sinon from 'sinon';
 
-import { NetworkInterfaceMonitor } from './../../src/network-interfaces/NetworkInterfaceMonitor';
-import * as data from './NetworkInterfaceData';
+import { getIPv4Addresses } from './../../src/network-interfaces/NetworkInterface';
+import * as mocks from './Mocks';
 
 chai.should();
 
-describe('when monitoring IPv4 addresses on network interfaces', function() {
+describe('NetworkInterface', function() {
+
     describe('#getIPv4Addresses', function() {
 
         let osStub: sinon.SinonStub;
-        const networkInterfaceMonitor = new NetworkInterfaceMonitor();
 
         afterEach(function() {
             osStub.restore();
@@ -20,10 +20,10 @@ describe('when monitoring IPv4 addresses on network interfaces', function() {
         it('should return addresses from one network interface', function() {
             // Arrange
             osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(data.NETWORK_INTERFACE_WITH_TWO_ADDRESSES);
+                .returns(mocks.NETWORK_INTERFACE_WITH_TWO_ADDRESSES);
 
             // Act
-            const addresses = networkInterfaceMonitor.getIPv4Addresses();
+            const addresses = getIPv4Addresses();
 
             // Assert
             addresses.should.be.eql(['1.1.1.1', '2.2.2.2']);
@@ -32,46 +32,46 @@ describe('when monitoring IPv4 addresses on network interfaces', function() {
         it('should return addresses from multiple network interfaces', function() {
             // Arrange
             osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(data.NETWORK_INTERFACES_WITH_TWO_ADDRESSES);
+                .returns(mocks.NETWORK_INTERFACES_WITH_TWO_ADDRESSES);
 
             // Act
-            const addresses = networkInterfaceMonitor.getIPv4Addresses();
+            const addresses = getIPv4Addresses();
 
             // Assert
             addresses.should.be.eql(['1.1.1.1', '2.2.2.2']);
         });
 
-        it('should return an empty sequence if only internal addresses exists', function() {
+        it('should not return internal addresses', function() {
             // Arrange
             osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(data.NETWORK_INTERFACES_WITH_INTERNAL_ADDRESSES);
+                .returns(mocks.NETWORK_INTERFACES_WITH_INTERNAL_ADDRESSES);
 
             // Act
-            const addresses = networkInterfaceMonitor.getIPv4Addresses();
+            const addresses = getIPv4Addresses();
 
             // Assert
             addresses.should.be.empty;
         });
 
-        it('should return an empty sequence if only IPv6 addresses exists', function() {
+        it('should not return IPv6 addresses', function() {
             // Arrange
             osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(data.NETWORK_INTERFACES_WITH_IPV6_ADDRESSES);
+                .returns(mocks.NETWORK_INTERFACES_WITH_IPV6_ADDRESSES);
 
             // Act
-            const addresses = networkInterfaceMonitor.getIPv4Addresses();
+            const addresses = getIPv4Addresses();
 
             // Assert
             addresses.should.be.empty;
         });
 
-        it('should return an empty sequence if no interfaces exists', function() {
+        it('should not fail on systems without network interfaces', function() {
             // Arrange
             osStub = sinon.stub(os, 'networkInterfaces')
-                .returns(data.NO_NETWORK_INTERFACES);
+                .returns(mocks.NO_NETWORK_INTERFACES);
 
             // Act
-            const addresses = networkInterfaceMonitor.getIPv4Addresses();
+            const addresses = getIPv4Addresses();
 
             // Assert
             addresses.should.be.empty;